(chore): rename .babelrc to babel.config.js
