--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    ["@babel/preset-env", {
-      "targets": {
-        "node": "6"
-      }
-    }]
-  ],
-  "plugins": [
-    "@babel/plugin-proposal-function-bind",
-    "@babel/plugin-proposal-export-default-from",
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
