+++ babel.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+    presets: [
+        ['@babel/preset-env', {
+            targets: {
+                node: 8
+            }
+        }]
+    ],
+    plugins: [
+        '@babel/plugin-proposal-function-bind',
+        '@babel/plugin-proposal-export-default-from',
+        '@babel/plugin-proposal-class-properties'
+    ]
+}
