+++ packages/wdio-config/src/utils.js
@@ -72,3 +72,49 @@ export function initialisePlugin (name, type) {
         )
     }
 }
+
+
+/**
+ * validates configurations based on default values
+ * @param  {Object} defaults  object describing all allowed properties
+ * @param  {Object} options   option to check agains
+ * @return {Object}           validated config enriched with default values
+ */
+export function validateConfig (defaults, options) {
+    const params = {}
+
+    for (const [name, expectedOption] of Object.entries(defaults)) {
+        /**
+         * check if options is given
+         */
+        if (typeof options[name] === 'undefined' && !expectedOption.default && expectedOption.required) {
+            throw new Error(`Required option "${name}" is missing`)
+        }
+
+        if (typeof options[name] === 'undefined' && expectedOption.default) {
+            params[name] = expectedOption.default
+        }
+
+        if (typeof options[name] !== 'undefined') {
+            if (typeof expectedOption.type === 'string' && typeof options[name] !== expectedOption.type) {
+                throw new Error(`Expected option "${name}" to be type of ${expectedOption.type} but was ${typeof options[name]}`)
+            }
+
+            if (typeof expectedOption.type === 'function') {
+                try {
+                    expectedOption.type(options[name])
+                } catch (e) {
+                    throw new Error(`Type check for option "${name}" failed: ${e.message}`)
+                }
+            }
+
+            if (expectedOption.match && !options[name].match(expectedOption.match)) {
+                throw new Error(`Option "${name}" doesn't match expected values: ${expectedOption.match}`)
+            }
+
+            params[name] = options[name]
+        }
+    }
+
+    return params
+}
